suppression et remboursement des joueur fait

+++ actions/item_delete.php
@@ -2,6 +2,8 @@
 
 require_once('../secret/connect_db.php');
 require_once('../classes/User.php');
+require_once('../classes/Item.php');
+require_once('../classes/Inventory.php');
 session_start();
 
 if(!isset($_SESSION['userId'])){
@@ -12,9 +14,9 @@ if(!isset($_SESSION['userId'])){
 $sql = "SELECT * FROM users WHERE id=?";
 $query = $db->prepare($sql);
 $query->execute(array($_SESSION['userId']));
-$user = $query->fetchObject('User');
+$curentUser = $query->fetchObject('User');
 
-if(!$user->getIsAdmin()){
+if(!$curentUser->getIsAdmin()){
     header('Location:../views/shop.php?status=danger&text=pas admin');
     exit;
 } 
@@ -24,22 +26,60 @@ if(!isset($_GET['id'])) {
     exit;
 }
 
-$sql = "DELETE FROM `inventories` WHERE `itemId`=?";
+$sql = "SELECT * FROM inventories WHERE itemId=?";
 $query = $db->prepare($sql);
-$is_success = $query->execute(array($_GET['id']));
+$query->execute(array($_GET['id']));
+$inventories = $query->fetchAll(PDO::FETCH_CLASS, "Inventory");
 
-if(!$is_success) {
-    header('Location: ../views/shop.php?status=danger&text=Une erreur est survenu, veillez réesseyer');
-    exit();
+if(!$inventories === false) {
+
+    foreach ($inventories as $inventory) {
+        $sql = "SELECT * FROM users WHERE id=?";
+        $query = $db->prepare($sql);
+        $query->execute(array($inventory->getUserId()));
+        $user = $query->fetchObject('User');
+
+        if($user === false) {
+            header('Location: ../views/shop.php?status=danger&text=Une erreur est survenu sur la table user, veillez réesseyer');
+            exit();
+        }
+
+        $sql = "SELECT * FROM items WHERE id=?";
+        $query = $db->prepare($sql);
+        $query->execute(array($inventory->getItemId()));
+        $item = $query->fetchObject('Item');
+
+        if($item === false) {
+            header('Location: ../views/shop.php?status=danger&text=Une erreur est survenu sur la table item, veillez réesseyer');
+            exit();
+        }
+
+        $sql = 'UPDATE `users` SET `money`=:money WHERE id=:id';
+        $query = $db->prepare($sql);
+        $is_success = $query->execute(array(":money"=>($user->getMoney()+$item->getPrice()), ":id"=>$user->getId()));
+
+        if(!$is_success) {
+            header('Location: ../views/shop.php?status=danger&text=Une erreur est survenu pour insertion money user, veillez réesseyer');
+            exit();
+        }
+
+        $sql = "DELETE FROM `inventories` WHERE `userId`=:userID and `itemId`=:itemId";
+        $query = $db->prepare($sql);
+        $is_success = $query->execute(array(':userID'=>$user->getId(),':itemId'=>$item->getId()));
+
+        if(!$is_success) {
+            header('Location: ../views/shop.php?status=danger&text=Une erreur est survenu pour la suppresion d\' un item dans l\'inventaire, veillez réesseyer');
+            exit();
+        }
+    }
 }
-
 
 $sql = "DELETE FROM `items` WHERE `id`=?";
 $query = $db->prepare($sql);
 $is_success = $query->execute(array($_GET['id']));
 
 if($is_success) {
-    header('Location: ../views/shop.php?status=success&text=l\'objet a était corectement ');
+    header('Location: ../views/shop.php?status=success&text=l\'objet a était corectement supprimer');
     exit();
 } else {
     header('Location: ../views/shop.php?status=danger&text=Une erreur est survenu, veillez réesseyer');
